+++ aiogram_bot/models/user_favorites.py
@@ -5,7 +5,7 @@ from aiogram_bot.models import Base
 class UserFavorites(Base):
     __tablename__ = 'User_favorites'
     id = sa.Column(sa.Integer, primary_key=True, autoincrement=True, nullable=False)
-    user_id = sa.Column(sa.Integer, nullable=False)
+    user_id = sa.Column(sa.BigInteger, nullable=False)
     resource = sa.Column(sa.Text, nullable=False)
     resource_type = sa.Column(sa.Text, nullable=False)
